refactor(#1619): flatten ticket-flow and Telegram re-export layers
Remove compatibility re-exports from runtime_helpers.py (resolve_run_reuse_policy,
select_active_or_paused_run, select_resumable_run) and telegram/state.py
(APPROVAL_MODES, TOPIC_ROOT, ThreadSummary, normalize_agent). Update all callers
to import from canonical modules (core.ticket_flow_operator, telegram.state_types).
