+++ src/codex_autorunner/flows/ticket_flow/runtime_helpers.py
@@ -29,12 +29,7 @@ from ...core.state_roots import resolve_repo_flows_db_path, resolve_repo_state_r
 from ...core.ticket_flow_operator import (
     PreflightCheckResult as TicketFlowInboxPreflight,
 )
-from ...core.ticket_flow_operator import (
-    ensure_flow_worker,
-    resolve_run_reuse_policy,  # noqa: F401
-    select_active_or_paused_run,  # noqa: F401
-    select_resumable_run,  # noqa: F401
-)
+from ...core.ticket_flow_operator import ensure_flow_worker
 from ...core.ticket_flow_operator import (
     ticket_flow_inbox_preflight as _ticket_flow_inbox_preflight,
 )
