+++ src/codex_autorunner/integrations/telegram/state.py
@@ -33,10 +33,6 @@ from .state_types import (
     parse_topic_key,
     topic_key,
 )
-from .state_types import APPROVAL_MODES as APPROVAL_MODES  # noqa: F401
-from .state_types import TOPIC_ROOT as TOPIC_ROOT  # noqa: F401
-from .state_types import ThreadSummary as ThreadSummary  # noqa: F401
-from .state_types import normalize_agent as normalize_agent  # noqa: F401
 
 logger = logging.getLogger("codex_autorunner.integrations.telegram.state")
